+++ src/PostgREST/RangeQuery.hs
@@ -48,9 +48,9 @@ rangeRequested headers = maybe allRange rangeParse $ lookup hRange headers
 
 restrictRange :: Maybe Integer -> NonnegRange -> NonnegRange
 restrictRange Nothing r = r
-restrictRange (Just limit) r =
-   rangeIntersection r $
-     Range BoundaryBelowAll (BoundaryAbove $ rangeOffset r + limit - 1)
+restrictRange (Just limit) r
+  | r == emptyRange = r
+  | otherwise = rangeIntersection r $ Range BoundaryBelowAll (BoundaryAbove $ rangeOffset r + limit - 1)
 
 rangeLimit :: NonnegRange -> Maybe Integer
 rangeLimit range =
@@ -77,8 +77,8 @@ rangeLeq n =
 
 rangeStatusHeader :: NonnegRange -> Int64 -> Maybe Int64 -> (Status, Header)
 rangeStatusHeader topLevelRange queryTotal tableTotal =
-  let lower = rangeOffset topLevelRange
-      upper = lower + toInteger queryTotal - 1
+  let lower = if topLevelRange /= emptyRange then rangeOffset topLevelRange else 0
+      upper = if topLevelRange /= emptyRange then lower + toInteger queryTotal - 1 else 0
       contentRange = contentRangeH lower upper (toInteger <$> tableTotal)
       status = rangeStatus lower upper (toInteger <$> tableTotal)
   in (status, contentRange)